Change the default mapping type for number from double to float
See https://www.elastic.co/guide/en/elasticsearch/reference/master/breaking_50_mapping_changes.html#_floating_points_use_literal_float_literal_instead_of_literal_double_literal
Closes #2003

+++ src/Nest/Mapping/Types/Core/Number/NumberAttribute.cs
@@ -1,5 +1,9 @@
 ﻿namespace Nest
 {
+	/// <summary>
+	/// Maps a property as a number type. If no type is specified,
+	/// the default type is float (single precision floating point).
+	/// </summary>
 	public class NumberAttribute : ElasticsearchPropertyAttributeBase, INumberProperty
 	{
 		INumberProperty Self => this;
@@ -23,6 +27,6 @@
 
 		public NumberAttribute(NumberType type) : base(type.GetStringValue()) { }
 		protected NumberAttribute(string type) : base(type) { }
-		public NumberAttribute() : base(NumberType.Double.GetStringValue()) { }
+		public NumberAttribute() : base(NumberType.Float.GetStringValue()) { }
 	}
 }
